feat:collect data system base on stm32f103 && fix(all):修复一些问题跟异常抛出

+++ src/ui/history/page_history.h
@@ -52,8 +52,8 @@ private:
     quint64 m_endTime = 0;
     quint64 m_startMaxTime = 0;
     quint64 m_endMaxTime = 0;
-    quint64 m_maxValue = 0;
-    quint64 m_minValue = 0;
+    qreal m_maxValue = 0;
+    qreal m_minValue = 0;
 
 };
